+++ src/components/basic_table/in_memory_table.test.js
@@ -670,5 +670,59 @@ describe('EuiInMemoryTable', () => {
 
       expect(component).toMatchSnapshot();
     });
+
+    test('onTableChange callback', () => {
+      const props = {
+        ...requiredProps,
+        items: [
+          { id: '1', name: 'name1' },
+          { id: '2', name: 'name2' },
+          { id: '3', name: 'name3' },
+          { id: '4', name: 'name4' },
+        ],
+        columns: [
+          {
+            field: 'name',
+            name: 'Name',
+            description: 'description',
+            sortable: true,
+          }
+        ],
+        sorting: true,
+        pagination: {
+          pageSizeOptions: [2, 4, 6],
+        },
+        onTableChange: jest.fn(),
+      };
+
+      const component = mount(
+        <EuiInMemoryTable {...props}/>
+      );
+
+      expect(props.onTableChange).toHaveBeenCalledTimes(0);
+      component.find('EuiPaginationButton[data-test-subj="pagination-button-1"]').simulate('click');
+      expect(props.onTableChange).toHaveBeenCalledTimes(1);
+      expect(props.onTableChange).toHaveBeenCalledWith({
+        sort: {},
+        page: {
+          index: 1,
+          size: 2,
+        },
+      });
+
+      props.onTableChange.mockClear();
+      component.find('[data-test-subj*="tableHeaderCell_name_0"] [data-test-subj="tableHeaderSortButton"]').simulate('click');
+      expect(props.onTableChange).toHaveBeenCalledTimes(1);
+      expect(props.onTableChange).toHaveBeenCalledWith({
+        sort: {
+          direction: 'asc',
+          field: 'name',
+        },
+        page: {
+          index: 0,
+          size: 2,
+        },
+      });
+    });
   });
 });